Fix a reference that re-anchored on a field that had not decayed
⚑ A REAL DEFECT, found by writing the boundary case the sweep said was
missing. A field correlating at EXACTLY the tracking threshold was counted as
having lost every one of its points, and re-anchored - abandoning every point
it could not measure on that frame, for nothing.

The cause is one this project has already diagnosed and fixed once, in
Recovery.cpp's isSeed(), which explains it at length: a correlation is a float
and a threshold typed into a spin box is a double, so widening the float to
compare turns 0.9f into 0.899999976, which sits just under a threshold of 0.9.
That fix never travelled the six lines to this comparison. Both sides are
narrowed to float here now, and the case that found it stands guard.

The other boundary beside it: the share is what must STILL be tracking, so a
field holding exactly that share is holding enough. Nine of ten points at a
required nine tenths does not re-anchor; eight does.

⚑ And a lost point reports NOTHING, not merely no displacement. The existing
case checked that it was not reported as a measurement and said nothing about
the three flags beside it, each a separate line free to go on saying
"measured". A point whose position in the current reference is unknown was
solved against the wrong pixels, so its strain, its noise floor and its match
conditioning are statements about somewhere it is not - and each reaches the
screen and the .vtu as an ordinary number.

In Recovery.cpp, the boundary of "may never make a point worse": an answer
correlating exactly as well as the one already there is not an improvement,
and accepting it would put the `recovered` mark - which is provenance - on a
point that gained nothing by it.

+++ src/core/ReferenceUpdate.cpp
@@ -44,7 +44,17 @@ bool fieldNeedsReanchor(const CorrelationResult &measured,
         if (!point.converged)
             continue;
         anyMeasurement++;
-        if (double(point.zncc) >= policy.znccThreshold)
+        // ⚑ COMPARED IN FLOAT, NOT IN DOUBLE, for the reason Recovery.cpp's
+        // isSeed() gives at length: a correlation is a float and a threshold
+        // typed into a spin box is a double, and widening the float to compare
+        // turns 0.9f into 0.899999976, which sits just under a threshold of
+        // 0.9. A field correlating at exactly the bar was therefore counted as
+        // having lost every one of its points, and re-anchored -- abandoning
+        // every point it could not measure on that frame, for nothing.
+        //
+        // The same fix as there, and it should have arrived with it: narrow
+        // the threshold once so both sides are in one representation.
+        if (point.zncc >= float(policy.znccThreshold))
             stillTracking++;
     }
